feat: implement deleteBook feature.

+++ controllers/book.js
@@ -1,6 +1,7 @@
 // Gestion de la logique de l'application du model Book
 
 const Book = require('../models/Book');
+const fs = require('fs');
 
 exports.getAllBooks = (req, res, next) => {
     Book.find()
@@ -99,9 +100,21 @@ exports.updateBook = (req, res, next) => {
 };
 
 exports.deleteBook = (req, res, next) => {
-    Book.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Book deleted successfully!' }))
-        .catch(error => res.status(400).json({ error }));
+    Book.findOne({ _id: req.params.id })
+        .then(book => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({ message: 'Not authorized' });
+            } else {
+                const filename = book.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Book.deleteOne({ _id: req.params.id })
+                        .then(() => res.status(200).json({ message: 'Book deleted successfully!' }))
+                        .catch(error => res.status(401).json({ error }));
+                });
+                
+            }             
+        })
+        .catch(error => res.status(500).json({ error }));    
 };
